Move internal protobufs into different folder
