+++ app/backend/src/couchers/email/__init__.py
@@ -7,7 +7,7 @@ from markdown2 import markdown
 from couchers import config
 from couchers.jobs.enqueue import queue_job
 from couchers.models import BackgroundJobType
-from pb import jobs_pb2
+from pb.internal import jobs_pb2
 
 loader = FileSystemLoader(Path(__file__).parent / ".." / ".." / ".." / "templates")
 env = Environment(loader=loader, trim_blocks=True)
